text_to_image_attngan/requirements.txt
cycler==0.10.0
decorator==4.4.1
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.4
nltk==3.4.5
numpy==1.18.0
pandas==0.24.2
Pillow==6.2.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.15.0
scipy==1.4.1
six==1.13.0
torch==1.3.1+cu92
torchvision==0.4.2+cu92

